Modify "if" statement to skip events based on NaN "PreferredTime" field
Previously would skip events if the "Type" field was empty.  I think this was
sufficient but it is possible that I missed events if there were actual events
which have no "Type" classification. Seems unlikely; but a search of the
unidentified events with this this updated sac2evt.m may find more matches.
Doubtful.