Correct reference time in xlabel
Previously stated the time was relative to "seisdate.B," the absolute
datetime of the first sample of the seismogram, when in actuality it
was relative to the reference time of the seismogram, "refdate".

To date this distinction is trivial given that h.B (offset from
refdate) is on the order of 1x10^-4 seconds for MERMAID data, but it
will be important for data with larger offset times.