Clarify how negative event depths are handled
Previously the header incorrectly stated that EQ.PreferredDepth was
overwritten to 0 km in the case of negative-depth events.  This is
incorrect. That field is and all other depth fields in the main struct
returned by irisFetch.Events is left unchanged. What is adjusted is
the srcDepth used to compute theoretical phase-arrival times in
arrivaltime.m, which will say 0 if the EQ.PreferredDepth is negative.