Remove call to external parcpci.m (parallel) in favor of local recursion
This diff looks huge but its not -- the only difference is that I
removed parcpci.m because it was finicky and hadn't really been
recursive for a while, anyway.  The diff looks large because I cleaned
up spacing, removed the main chunk of code from an 'else' statement by
adding 'return' after the recursion.