Return empty updated EQ if no phase arrivals in time window of seismogram
This change was made to mirror the output of sac2evt.m.  Previously
this function returned the 'bare' EQ structure with .TaupTimes = [].

Also added model and phase inputs (and corrected those fields in the
updated EQ), as well as set .Params = NaN in the updated event after
realizing the .Params list no longer applies after update.