Adjust taper so that its length is a consistent proportion of the requested window length
Now only a portion of the data are tapered, not the entire input, so that the
effect is consistent regardless if the input time series is 200 s or 60 m in
length.  Also, add flags warning if taper failed (too long for time series), or
if there are suspected filler/missing values (implied by two contiguous datum
equaling exactly 0).