Verify no reviewed .evt exists before logging fail
The current "already matched" list relies on *raw.evt because that is
how I defined a "matched" event, regardless of its review status. If I
swapped the list to use the reviewed events we may have lots of
unnecessary redo attempts in cases when a raw.evt but not a rev.evt
exists.