Refactor and swap resolution input from m to deg
Also massively refactor code for readability.  Moved original
fresnelgrid.m to $OMNIA/atm. That one was faster, and I think could
accomplish the same thing...I just need to work on it a little more?  I
think this version has a slightly fatter (closer to reality) Fresnel
zone for more of the path, due to how I was indexing arcs and whatnot.
Maybe just forget about that one; this seems to work.